Updated
This might be a poor update.  The last firmware update to my TED5000
borked the posting of usage data.  Before I tracked the problem down to
the TED I had ripped this code open a couple of times.
When I get a fix from the folks at http://www.theenergydetective.com/ I
will have another update to this file